// src/types/cart.ts
export interface CurrencyInfo {
  currency_code: CurrencyCode;
  currency_symbol: string;
  currency_minor_unit: number;
  currency_decimal_separator: string;
  currency_thousand_separator: string;
  currency_prefix: string;
  currency_suffix: string;
}

export interface CartTotalsItem extends CurrencyInfo {
  total_discount: string;
  total_discount_tax: string;
}

export interface CartCouponItem {
  code: string;
  label: string;
  discount_type: string;
  totals: CartTotalsItem;
}

export interface FirstNameLastName {
  first_name: string;
  last_name: string;
}

export interface BaseAddress {
  address_1: string;
  address_2: string;
  city: string;
  state: string;
  postcode: string;
  country: string;
}

export type AddressFieldsForShippingRates = Array<Exclude<keyof BaseAddress, 'address_1' | 'address_2'>>;

export interface CartShippingPackageShippingRate extends CurrencyInfo {
  rate_id: string;
  name: string;
  description: string;
  delivery_time: string;
  price: string;
  taxes: string;
  instance_id: number;
  method_id: string;
  meta_data: Array<MetaKeyValue>;
  selected: boolean;
}

export interface CartShippingRate {
  package_id: string | number;
  name: string;
  destination: BaseAddress;
  items: Array<ShippingRateItem>;
  shipping_rates: Array<CartShippingPackageShippingRate>;
}